requirements.txt
streamlit>=1.36.0,<2.0.0

pydantic>=2.5.0,<3.0.0
python-dateutil>=2.8.2

SQLAlchemy>=2.0.25,<3.0.0
bcrypt>=4.0.0,<5.0.0

groq>=0.9.0

firecrawl-py>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

python-dotenv>=1.0.0

pandas>=2.1.0
PyYAML>=6.0.1

PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
rapidfuzz>=3.0.0
